requirements.txt
ijson
matplotlib
numpy
pandas
scipy
tqdm
pyyaml